+++ lib/recurly/transaction_list.php
@@ -17,6 +17,13 @@ class Recurly_TransactionList extends Recurly_Pager
     return $list;
   }
   
+  public static function getForAccount($accountCode, $params = null, $client = null)
+  {
+    $list = new Recurly_TransactionList(Recurly_Client::PATH_ACCOUNTS . '/' . rawurlencode($accountCode) . Recurly_Client::PATH_TRANSACTIONS, $client);
+    $list->_loadFrom(Recurly_Client::PATH_ACCOUNTS . '/' . rawurlencode($accountCode) . Recurly_Client::PATH_TRANSACTIONS, $params);
+    return $list;
+  }
+  
   protected function getNodeName() {
     return 'transactions';
   }